requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
black==24.3.0
celery==5.3.6
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cryptography==42.0.5
defusedxml==0.7.1
dj-database-url==2.1.0
Django==5.0.4
django-filter==24.2
django-rest-auth==0.9.5
django-templated-mail==1.1.1
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
djoser==2.2.2
elastic-transport==8.13.0
elasticsearch==7.17.6
Faker==24.9.0
gunicorn==21.2.0
idna==3.7
iniconfig==2.0.0
isort==5.13.2
kombu==5.3.6
mypy-extensions==1.0.0
oauthlib==3.2.2
packaging==24.0
pathspec==0.12.1
pillow==10.3.0
platformdirs==4.2.0
pluggy==1.4.0
prompt-toolkit==3.0.43
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
pytest==8.1.1
pytest-django==4.8.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python3-openid==3.2.0
redis==5.0.3
requests==2.31.0
requests-oauthlib==2.0.0
six==1.16.0
social-auth-app-django==5.4.0
social-auth-core==4.5.3
sqlparse==0.4.4
transliterate==1.10.2
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.18
vine==5.1.0
wcwidth==0.2.13
